BaseZz

Documentation for BaseZz.

BaseZz.fastextrema

BaseZz.isnotnumber

BaseZz.isnumber

BaseZz.fastextrema — Function

fastextrema(x) -> (mini, maxi)

Compute the minimum and maximum values of an input array x and return them as a tuple.

x can be an array of Real numbers, Gray colorants, or multi-channel colorants Colorant{T,N}, where T is a Real type. If x is an array of AbstractFloat elements, values such as NaN, missing, Inf, or -Inf are automatically excluded from the computation. Additionally, x can be a Skip object from the Skipper.jl package, created using the skip or keep functions.

Examples

julia> # Example with an array of real numbers
julia> data = [1.0, 2.0, 3.0, 4.0, 5.0]
julia> mini, maxi = fastextrema(data)
julia> println("Minimum: ", mini, ", Maximum: ", maxi)
Minimum: 1.0, Maximum: 5.0

julia> # Example with a multi-channel image
julia> img = rand(RGB{N0f8}, 10, 10) # Random RGB image
julia> mini, maxi = fastextrema(img)
julia> println("Minimum: ", mini, ", Maximum: ", maxi)
Minimum: RGB{N0f8}(0.004, 0.004, 0.0), Maximum: RGB{N0f8}(0.996, 0.992, 0.996)

julia> # Example with an 8-bit grayscale image, ignoring all values below 0.5
julia> img = rand(Gray{N0f8}, 10, 10)
julia> mini, maxi = skip(x -> x < 0.5, img) |> fastextrema
julia> println("Minimum: ", mini, ", Maximum: ", maxi)
Minimum: Gray{N0f8}(0.51), Maximum: Gray{N0f8}(0.984)

Note

Microbenchmarks indicate that fastextrema is about twice as fast as Base.extrema for floating-point numbers, but shows no significant benefit for integers. The advantage of fastextrema lies in its support for Skipper.jl. However, it is less versatile than Base.extrema as it lacks the dims argument and does not support generic iterable collections. Despite this, fastextrema should remain sufficient for image processing tasks.

BaseZz.isnotnumber — Method

isnotnumber(x) -> Bool

Check if x is not a valid number.

Returns true if x is not finite, is NaN, or missing. Otherwise, returns false.

Examples

julia> isnotnumber(NaN)
true

julia> isnotnumber(Inf)
true

julia> isnotnumber(missing)
true

julia> isnotnumber(5.0)
false

See also: isnumber

BaseZz.isnumber — Method

isnumber(x) -> Bool

Check if x is a valid number.

Returns true if x is finite and not NaN or missing. Otherwise, returns false.

Examples

julia> isnumber(5.0)
true

julia> isnumber(NaN)
false

julia> isnumber(-Inf)
false

julia> isnumber(missing)
false

See also: isnotnumber